requirements.txt
opencv-python
mediapipe
numpy
scikit-learn
joblib
pandas
pygame
PyOpenGL
